ui: anduril: Add support for ultra-low mode
Port forward the patch by tocirahl and add a config to enable it at runtime,
which allows level 1 to be set to use the default level 1, red aux, leakage
current in run mode, or leakage current in standby mode. Note that this
feature only applies to level 1, so the floor may need to be configured low
enough to use it.

This config needs to be enabled in the hwdef in supported lights. Only a few
linear driven lights can use this feature.